update storybook and move component to correct spot
