--- packages/manager/src/components/Accordion/index.ts
@@ -1,5 +1,0 @@
-import Accordion, { Props as _AccordionProps } from './Accordion';
-/* tslint:disable */
-export interface AccordionProps extends _AccordionProps {}
-
-export default Accordion;
