--- packages/manager/src/components/core/Accordion.ts
@@ -1,8 +1,0 @@
-import Accordion, {
-  AccordionProps as _AccordionProps,
-} from '@mui/material/Accordion';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface AccordionProps extends _AccordionProps {}
-
-export default Accordion;
