--- packages/manager/src/components/core/AccordionDetails.ts
@@ -1,8 +1,0 @@
-import AccordionDetails, {
-  AccordionDetailsProps as _AccordionDetailsProps,
-} from '@mui/material/AccordionDetails';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface AccordionDetailsProps extends _AccordionDetailsProps {}
-
-export default AccordionDetails;
