--- packages/manager/src/components/core/AccordionSummary.ts
@@ -1,8 +1,0 @@
-import AccordionSummary, {
-  AccordionSummaryProps as _AccordionSummaryProps,
-} from '@mui/material/AccordionSummary';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface AccordionSummaryProps extends _AccordionSummaryProps {}
-
-export default AccordionSummary;
