+++ packages/manager/src/features/Account/AutoBackups.tsx
@@ -1,7 +1,7 @@
 import { makeStyles } from 'tss-react/mui';
 import { Theme } from '@mui/material/styles';
 import * as React from 'react';
-import Accordion from 'src/components/Accordion';
+import { Accordion } from 'src/components/Accordion';
 import FormControlLabel from 'src/components/core/FormControlLabel';
 import Grid from '@mui/material/Unstable_Grid2';
 import { Notice } from 'src/components/Notice/Notice';
