+++ packages/manager/src/features/Account/CloseAccountSetting.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Accordion from 'src/components/Accordion';
+import { Accordion } from 'src/components/Accordion';
 import Button from 'src/components/Button';
 import Grid from '@mui/material/Unstable_Grid2';
 import CloseAccountDialog from './CloseAccountDialog';
