+++ packages/manager/src/features/Account/NetworkHelper.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import FormControlLabel from 'src/components/core/FormControlLabel';
 import Typography from 'src/components/core/Typography';
-import Accordion from 'src/components/Accordion';
+import { Accordion } from 'src/components/Accordion';
 import Grid from '@mui/material/Unstable_Grid2';
 import { Toggle } from 'src/components/Toggle';
