+++ packages/manager/src/features/NodeBalancers/NodeBalancerCreate.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import ActionsPanel from 'src/components/ActionsPanel';
 import Button from 'src/components/Button';
 import Box from 'src/components/core/Box';
-import Accordion from 'src/components/Accordion';
+import { Accordion } from 'src/components/Accordion';
 import Paper from 'src/components/core/Paper';
 import LandingHeader from 'src/components/LandingHeader';
 import { TextField } from 'src/components/TextField';
