+++ packages/manager/src/features/Support/SupportTickets/SupportTicketDrawer.tsx
@@ -7,7 +7,7 @@ import * as Bluebird from 'bluebird';
 import { update } from 'ramda';
 import * as React from 'react';
 import { compose as recompose } from 'recompose';
-import Accordion from 'src/components/Accordion';
+import { Accordion } from 'src/components/Accordion';
 import { StyledActionPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import Button from 'src/components/Button';
 import FormHelperText from 'src/components/core/FormHelperText';
